chore: add test coverage
